Make sure offset_x/y is even before CropAndScale()
CropAndScale() makes offset_x and offset_y even if the U,V planes are
subsampled. This may result in the update rect being off by one. To
prevent that, pass even offset_x and offset_y to CropAndScale() if they
would be made even under the hood.

Note: The code was originally added in
https://webrtc-review.googlesource.com/c/src/+/123230.

Change-Id: I4a70d460702f03dee72a5b8292cb25766c3e6aca